requirements.txt
streamlit
google-generativeai
ics